requirements.txt
python-dotenv
# lock telegram bot version
python-telegram-bot==22.5
Pillow
matplotlib
requests
maidenhead
beautifulsoup4
pyzbar
tmdbv3api
# pilmoji broken: https://github.com/jay3332/pilmoji/pull/40
emoji==2.11.0
cairosvg
googletrans-py
dice
mcstatus
pilmoji
